增加新浪财经的数据下载

+++ src/main/java/com/liberty/system/web/KlineController.java
@@ -19,7 +19,8 @@ import com.liberty.system.model.Kline;
 import com.liberty.system.model.Shape;
 import com.liberty.system.model.Stroke;
 import com.liberty.system.service.DownLoader;
-import com.liberty.system.service.HxDownLoader;
+import com.liberty.system.service.impl.HxDownLoader;
+import com.liberty.system.service.impl.XlDownLoader;
 
 public class KlineController extends BaseController {
 	private static final Map<String, String> klineTypeNumberMap;
@@ -62,104 +63,32 @@ public class KlineController extends BaseController {
 		// System.err.println(shapes.get(shapes.size() - 1).getMax());
 		// System.out.println(last.equals(shapes.get(shapes.size() - 1)));
 
-		long date = Long.valueOf("1531789177122");
-		date = date - 8 * 60 * 60 * 1000;
-		System.out.println(new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(new Date(date)));
+		// long date = Long.valueOf("1531789177122");
+		// System.out.println("1111" + new Date().getTime());
+		// date = date - 8 * 60 * 60 * 1000;
+		// System.out.println(new SimpleDateFormat("yyyy-MM-dd
+		// HH:mm:ss").format(new Date(date)));
+
+//		String[] split = "12|34|56|".split("\\|");
+//
+//		System.out.println(split[1]);
+		
 
 	}
 
 	/**
-	 * 新浪 K线数据下载
+	 * K线数据下载
 	 * 
 	 */
-	public void xlDownloadData() {
-		String response = "";
-		String dataUrl = "https://ex.sina.com.cn/forex/api/jsonp.php/var%20_fx_seurusd_1_1531791999571=/NewForexService.getMinKline?symbol=fx_seurusd&scale=1&datalen=1440";
-		String dataUrl1 = "https://ex.sina.com.cn/forex/api/jsonp.php/var%20_fx_seurusd2018_7_17=/NewForexService.getDayKLine?symbol=fx_seurusd&_=2018_7_17";
-		Map<String, String> params = new HashMap<String, String>();
+	public void downloadData() {
+		DownLoader downLoader = new XlDownLoader();
 		List<Currency> listAll = Currency.dao.listAll();
-
-		Date now = new Date();
 
 		String sql = "select * from dictionary where type='klineType'";
 		List<Record> klineTyep = Db.find(sql);
 		for (Record record : klineTyep) {
 			// ================测试,只取五分钟k线
-			if (!record.getStr("key").equals("1")) {
-				continue;
-			}
-			// ================
-			Map<String, List<Kline>> klineMap = new HashMap<String, List<Kline>>();
-			Map<String, Kline> lastKlineMap = new HashMap<String, Kline>();
-			params.put("type", record.getStr("key"));// K线级别
-
-			for (Currency currency : listAll) {
-				// ===========测试,只取eurusd
-				// if (!currency.getCode().equals("EURUSD")) {
-				// continue;
-				// }
-				// ===========
-				List<Kline> klineList = new ArrayList<Kline>();
-				params.put("code", "FOREX" + currency.getCode());// 设置code参数
-				// 取出最后两条数据,最新的一条数据可能随时变化,新增数据时此条记录先删除
-				List<Kline> lastTwo = Kline.dao.getLastByCode(currency.getCode(), record.getStr("key"));
-				// 设置number参数
-				if (lastTwo == null || lastTwo.size() == 0) {
-					params.put("number", klineTypeNumberMap.get(record.getStr("key")) == null ? "-1000"
-							: klineTypeNumberMap.get(record.getStr("key")));
-				} else {
-					lastTwo.get(0).delete();
-					lastKlineMap.put(currency.getCode(), lastTwo.get(1));
-					Date lastDate = lastTwo.get(1).getDate();
-					long between = DateUtil.getNumberBetween(DateUtil.getNextDay(now), lastDate,
-							klineTypeBetweenMap.get(record.getStr("key")));
-					String number = String.valueOf(between);
-					params.put("number", "-" + number);
-					// params.put("number", "-" + "10");//测试
-				}
-				response = HTTPUtils.http(dataUrl, params, "get");
-				response = response.substring(response.indexOf("{"));
-				response = response.substring(0, response.lastIndexOf("}") + 1);
-				Map<String, Object> responseMap = JSON.parseObject(response, Map.class);
-				Object object = responseMap.get("Data");
-				JSONArray parseArray = JSON.parseArray(object.toString());
-				Object startDate = parseArray.get(1);// 开始时间
-				Object endDate = parseArray.get(2);// 结束时间
-				Object priceMul = parseArray.get(4);// 价格倍数
-
-				JSONArray dataArray = JSON.parseArray(parseArray.get(0).toString());// 数据数组
-				for (Object object2 : dataArray) {
-					Kline kline = new Kline();
-					JSONArray parseArray2 = JSON.parseArray(object2.toString());
-					kline.setDate(DateUtil.strDate(parseArray2.get(0).toString(), "yyyyMMddHHmmss"));
-					kline.setMax(Double.valueOf(parseArray2.get(4).toString()) / Double.valueOf(priceMul.toString()));
-					kline.setMin(Double.valueOf(parseArray2.get(5).toString()) / Double.valueOf(priceMul.toString()));
-					kline.setCurrencyId(currency.getId());
-					kline.setType(record.getStr("key"));
-
-					klineList.add(kline);
-				}
-
-				klineMap.put(currency.getCode(), klineList);
-			}
-			Kline.dao.saveMany(klineMap, lastKlineMap);
-		}
-		renderText(response);
-	}
-
-	/**
-	 * 和讯 K线数据下载
-	 * 
-	 */
-	public void hexunDownloadData() {
-		DownLoader downLoader = new HxDownLoader();
-		List<Currency> listAll = Currency.dao.listAll();
-
-		String sql = "select * from dictionary where type='hxKlineType'";
-		List<Record> klineTyep = Db.find(sql);
-		for (Record record : klineTyep) {
-			// ================测试,只取五分钟k线
-			if (!record.getStr("key").equals("1")) {
+			if (!record.getStr("key").equals("7")) {
 				continue;
 			}
 			// ================
@@ -168,27 +97,28 @@ public class KlineController extends BaseController {
 
 			for (Currency currency : listAll) {
 				// ===========测试,只取eurusd
-				// if (!currency.getCode().equals("EURUSD")) {
-				// continue;
-				// }
+				if (!currency.getCode().equals("EURUSD")) {
+					continue;
+				}
 				// ===========
 				// 取出最后两条数据,最新的一条数据可能随时变化,新增数据时此条记录先删除
 				List<Kline> lastTwo = Kline.dao.getLastByCode(currency.getCode(), record.getStr("key"));
-				
-				// 设置number参数
-				Kline lastKline=null;
+
+				Kline lastKline = null;
 				if (lastTwo == null || lastTwo.size() <= 1) {
-					if(lastTwo.size()==1){
+					if (lastTwo.size() == 1) {
 						lastTwo.get(0).delete();
 					}
 				} else {
 					lastTwo.get(0).delete();
-					lastKline=lastTwo.get(1);
+					lastKline = lastTwo.get(1);
 					lastKlineMap.put(currency.getCode(), lastKline);
-					Date lastDate = lastTwo.get(1).getDate();
 				}
-
-				List<Kline> klineList = downLoader.downLoad(currency.getCode(), record.getStr("key"), "get",lastKline);
+				List<Kline> klineList = null;
+				klineList = downLoader.downLoad(currency.getCode(), record.getStr("key"), "get", lastKline);
+				if(klineList==null || klineList.size()==0){
+					continue;
+				}
 				for (Kline kline : klineList) {
 					kline.setCurrencyId(currency.getId());
 					kline.setType(record.getStr("key"));
@@ -218,7 +148,7 @@ public class KlineController extends BaseController {
 
 		if (lastStroke == null) {
 			// 查询所有的K线
-			List<Kline> klines = Kline.dao.listAllByCode("EURUSD", "1");
+			List<Kline> klines = Kline.dao.listAllByCode("EURUSD", "7");
 			// 处理K线的包含关系
 			List<Kline> handleInclude = handleInclude(klines, lastStroke);
 			// 生成笔
